backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
